+++ Resume/Areas/Admin/Controllers/ExperienceController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Resume.Business.Control;
 using Resume.Business.Tools;
 using Resume.Models.Context;
 using Resume.Models.Entities;
@@ -15,7 +16,7 @@ namespace Resume.Areas.Admin.Controllers
     public class ExperienceController : Controller
     {
         private readonly ResumeContext _context;
-
+        CurrentUser currentUser = new CurrentUser();
         public ExperienceController(ResumeContext context)
         {
             _context = context;
@@ -23,13 +24,31 @@ namespace Resume.Areas.Admin.Controllers
 
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Experiences.ToListAsync());
+            bool roleStatus = RoleChecker.AuthorizeRoles(_context, currentUser.FindUser(_context, User.Identity.Name), "Experience", "Index");
+            if (roleStatus)
+            {
+                return View(await _context.Experiences.ToListAsync());
+            }
+            else
+            {
+                return Redirect("/Account/Denied");
+            }
+
         }
 
 
         public IActionResult Create()
         {
-            return View();
+            bool roleStatus = RoleChecker.AuthorizeRoles(_context, currentUser.FindUser(_context, User.Identity.Name), "Experience", "Create");
+            if (roleStatus)
+            {
+                return View();
+            }
+            else
+            {
+                return Redirect("/Account/Denied");
+            }
+
         }
 
 
@@ -37,30 +56,48 @@ namespace Resume.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(Experience experience)
         {
-            if (ModelState.IsValid)
+            bool roleStatus = RoleChecker.AuthorizeRoles(_context, currentUser.FindUser(_context, User.Identity.Name), "Experience", "Create");
+            if (roleStatus)
             {
-                _context.Add(experience);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(experience);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                return View(experience);
             }
-            return View(experience);
+            else
+            {
+                return Redirect("/Account/Denied");
+            }
+
         }
 
 
         public async Task<IActionResult> Edit(string id)
         {
-            if (id == null || IDAncryption.Decrypt(id) == "NotFound")
+            bool roleStatus = RoleChecker.AuthorizeRoles(_context, currentUser.FindUser(_context, User.Identity.Name), "Experience", "Edit");
+            if (roleStatus)
             {
-                return NotFound();
-            }
-            int dID = Convert.ToInt32(IDAncryption.Decrypt(id));
+                if (id == null || IDAncryption.Decrypt(id) == "NotFound")
+                {
+                    return NotFound();
+                }
+                int dID = Convert.ToInt32(IDAncryption.Decrypt(id));
 
-            var experience = await _context.Experiences.FindAsync(dID);
-            if (experience == null)
-            {
-                return NotFound();
+                var experience = await _context.Experiences.FindAsync(dID);
+                if (experience == null)
+                {
+                    return NotFound();
+                }
+                return View(experience);
             }
-            return View(experience);
+            else
+            {
+                return Redirect("/Account/Denied");
+            }
+
         }
 
 
@@ -68,51 +105,69 @@ namespace Resume.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, Experience experience)
         {
-            if (id != experience.ID)
+            bool roleStatus = RoleChecker.AuthorizeRoles(_context, currentUser.FindUser(_context, User.Identity.Name), "Experience", "Edit");
+            if (roleStatus)
             {
-                return NotFound();
+                if (id != experience.ID)
+                {
+                    return NotFound();
+                }
+
+                if (ModelState.IsValid)
+                {
+                    try
+                    {
+                        _context.Update(experience);
+                        await _context.SaveChangesAsync();
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!ExperienceExists(experience.ID))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
+                    }
+                    return RedirectToAction(nameof(Index));
+                }
+                return View(experience);
+            }
+            else
+            {
+                return Redirect("/Account/Denied");
             }
 
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(experience);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ExperienceExists(experience.ID))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(experience);
         }
 
 
         public async Task<IActionResult> Delete(string id)
         {
-            if (id == null || IDAncryption.Decrypt(id) == "NotFound")
+            bool roleStatus = RoleChecker.AuthorizeRoles(_context, currentUser.FindUser(_context, User.Identity.Name), "Experience", "Delete");
+            if (roleStatus)
             {
-                return NotFound();
-            }
-            int dID = Convert.ToInt32(IDAncryption.Decrypt(id));
+                if (id == null || IDAncryption.Decrypt(id) == "NotFound")
+                {
+                    return NotFound();
+                }
+                int dID = Convert.ToInt32(IDAncryption.Decrypt(id));
 
-            var experience = await _context.Experiences
-                .FirstOrDefaultAsync(m => m.ID == dID);
-            if (experience == null)
+                var experience = await _context.Experiences
+                    .FirstOrDefaultAsync(m => m.ID == dID);
+                if (experience == null)
+                {
+                    return NotFound();
+                }
+
+                return View(experience);
+            }
+            else
             {
-                return NotFound();
+                return Redirect("/Account/Denied");
             }
 
-            return View(experience);
         }
 
 
@@ -120,10 +175,19 @@ namespace Resume.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var experience = await _context.Experiences.FindAsync(id);
-            _context.Experiences.Remove(experience);
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            bool roleStatus = RoleChecker.AuthorizeRoles(_context, currentUser.FindUser(_context, User.Identity.Name), "Experience", "Delete");
+            if (roleStatus)
+            {
+                var experience = await _context.Experiences.FindAsync(id);
+                _context.Experiences.Remove(experience);
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
+            }
+            else
+            {
+                return Redirect("/Account/Denied");
+            }
+
         }
 
         private bool ExperienceExists(int id)
